Allowing different location and filename for the configuration file
By just changing the order of reading command arguements and configuration file, it is possible to specify a configuration file in a different location than the default one. This can be very useful when you need to e.g. setup a systemd timer and you might want to place the script and configuration file in different locations